+++ src/main/java/nextstep/courses/domain/coverimage/ImageFileSize.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.coverimage;
 
 public class ImageFileSize {
   private static final Long FILE_MAX_SIZE = 1024L*1024L;
